requirements.txt
absl-py==2.1.0
altgraph==0.17.4
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
autopep8==2.3.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Django==5.0.7
django-ckeditor==6.7.2
django-extensions==3.2.3
django-js-asset==3.0.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
Faker==33.1.0
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
fuzzywuzzy==0.18.0
gprof2dot==2024.6.6
idna==3.7
inflection==0.5.1
intel-openmp==2021.4.0
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
kombu==5.3.7
Levenshtein==0.26.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mongoengine==0.29.1
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.23.5
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pefile==2023.2.7
pillow==10.2.0
prompt_toolkit==3.0.47
protobuf==4.21.2
psutil==6.0.0
pure-eval==0.2.2
pycodestyle==2.12.1
Pygments==2.18.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
RapidFuzz==3.10.0
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.22.3
scikit-surprise==1.1.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
tomli==2.1.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.7.0
django-cors-headers==3.14.0
